// upvertice-ui/src/app/services/models/supplier-offer-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SponsorAdResponse } from '../models/sponsor-ad-response';
export interface SupplierOfferResponse {
  description?: string;
  endDate?: string;
  id?: number;
  imageUrl?: string;
  price?: number;
  quantityAvailable?: number;
  sponsorAds?: Array<SponsorAdResponse>;
  startDate?: string;
  status?: 'COMING_SOON' | 'AVAILABLE' | 'CLOSED';
  title?: string;
}